refactor(Gax): retry vs. polling policies
Make room to implement both retry and polling policies using the same
basic types, such as `LimitedAttemptCount<P>`.
